// src/app/interfaces/chat.ts
export interface Chat {
    //Nome do destinatário
    nome?: string;

    //Mensagens do remetente
    msgx1?: string;
    msgx2?: string;
    msgx3?: string;
    msgx4?: string;
    msgx5?: string;
    msgx6?: string;
    msgx7?: string;
    msgx8?: string;
    msgx9?: string;
    msgx10?: string;
    msgx11?: string;
    msgx12?: string;
    msgx13?: string;
    msgx14?: string;
    msgx15?: string;
    msgx16?: string;
    msgx17?: string;
    msgx18?: string;
    msgx19?: string;
    msgx20?: string;

    //Mensagens do destinatário
    msgy1?: string;
    msgy2?: string;
    msgy3?: string;
    msgy4?: string;
    msgy5?: string;
    msgy6?: string;
    msgy7?: string;
    msgy8?: string;
    msgy9?: string;
    msgy10?: string;
    msgy11?: string;
    msgy12?: string;
    msgy13?: string;
    msgy14?: string;
    msgy15?: string;
    msgy16?: string;
    msgy17?: string;
    msgy18?: string;
    msgy19?: string;
    msgy20?: string;
}